Move and rename useFreeAdCredit file.
